requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
arrow==1.2.3
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
chatgptpy==1.0.8
click==8.1.3
colorama==0.4.6
cssselect2==0.7.0
EmbyPy==0.6.6.4
exceptiongroup==1.1.2
fastapi==0.99.1
frozenlist==1.3.3
h11==0.14.0
idna==3.4
lxml==4.9.3
markdownify==0.11.6
multidict==6.0.4
numpy==1.25.0
oauthlib==3.2.2
openai==0.27.8
outcome==1.2.0
overrides==7.3.1
Pillow==10.0.0
PlexAPI==4.14.0
pyarr==5.1.0
pyChatGPT==0.4.3.3
pydantic==1.10.10
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.10.0
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
starlette==0.27.0
svglib==1.5.1
tinycss2==1.2.1
tls-client==0.2.1
tqdm==4.65.0
trakt==3.4.0
trio==0.22.2
trio-websocket==0.10.3
types-requests==2.31.0.1
types-urllib3==1.26.25.13
typing_extensions==4.7.1
undetected-chromedriver==3.5.0
urllib3==2.0.3
utils==1.0.1
uvicorn==0.22.0
webencodings==0.5.1
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.2

pymongo~=4.4.1
